Crate libbz2_rs_sys

A drop-in compatible rust implementation of bzip2

Structs

bz_stream

The current stream state.

Constants

BZ_CONFIG_ERROR

BZ_DATA_ERROR

BZ_DATA_ERROR_MAGIC

BZ_FINISH

BZ_FINISH_OK

BZ_FLUSH

BZ_FLUSH_OK

BZ_IO_ERROR

BZ_MAX_UNUSED

BZ_MEM_ERROR

BZ_OK

BZ_OUTBUFF_FULL

BZ_PARAM_ERROR

BZ_RUN

BZ_RUN_OK

BZ_SEQUENCE_ERROR

BZ_STREAM_END

BZ_UNEXPECTED_EOF

Functions

BZ2_bzBuffToBuffCompress⚠

Compress the input data into the destination buffer.

BZ2_bzBuffToBuffDecompress⚠

Decompress the input data into the destination buffer.

BZ2_bzCompress⚠

Compresses as much data as possible, and stops when the input buffer becomes empty or the output buffer becomes full.

BZ2_bzCompressEnd⚠

Deallocates all dynamically allocated data structures for this stream.

BZ2_bzCompressInit⚠

Prepares the stream for compression.

BZ2_bzDecompress⚠

Decompresses as much data as possible, and stops when the input buffer becomes empty or the output buffer becomes full.

BZ2_bzDecompressEnd⚠

Deallocates all dynamically allocated data structures for this stream.

BZ2_bzDecompressInit⚠

Prepares the stream for decompression.