libbz2_rs_sys

Function BZ2_bzCompress

pub unsafe extern "C" fn BZ2_bzCompress(
    strm: *mut bz_stream,
    action: c_int,
) -> c_int

Compresses as much data as possible, and stops when the input buffer becomes empty or the output buffer becomes full.

Returns

BZ_SEQUENCE_ERROR if called on an invalid stream, e.g.

before BZ2_bzCompressInit

after BZ2_bzCompressEnd

BZ_PARAM_ERROR if any of

strm.is_null()

strm.s.is_null()

action is not one of BZ_RUN, BZ_FLUSH or BZ_FINISH

BZ_RUN_OK successfully compressed, but ran out of input or output space

BZ_FLUSH_OK not all compressed data has been written to the output yet

BZ_FINISH_OK if all input has been read but not all output has been written to the output buffer yet

BZ_STREAM_END if all input has been read all output has been written to the output buffer

Safety

Either

strm is NULL

strm satisfies the requirements of &mut *strm and was initialized with BZ2_bzCompressInit

Either

strm.next_in is NULL and strm.avail_in is 0

strm.next_in is readable for strm.avail_in bytes

Either

strm.next_out is NULL and strm.avail_out is 0

strm.next_out is writable for strm.avail_out bytes